Add files via upload
Sorry I didn't have time to change the images like you asked. Please feel free to do it. Please have a look at the scale_misfit variable. The one in your code was returning an error in my version. Just make sure it is ok to do it like this.